refactor: lift the catalog search out of the agent factory
_search_catalog_impl was a 943-line closure inside _create_agent, which made it
unreachable from a test and impossible to read without also reading the agent
that built it. It captured six things from that scope; SearchContext names them,
so the search is now an ordinary function in its own module.

Verified mechanically rather than by inspection: each moved definition's body is
AST-identical to its original once the six captures are rewritten as context
attributes, and each signature is the original one with ctx prepended. The check
caught a real bug on the first attempt -- a word-boundary rename had rewritten
"scope" and "state" inside model-facing prompt strings.

Four grounding notes move to turn_support because both modules now read them.
