+++ tests/unit/catalog_retriever/test_utils.py
@@ -14,7 +14,6 @@ from __future__ import annotations
 import base64
 import io
 from typing import Any
-from unittest.mock import patch
 
 import pytest
 from PIL import Image
